# Faster H264Reader