Fixes #505 

The following is benchmark with same code in #505

```
$ cargo run --example nal-parser -- ~/test.h264          
    Finished dev [unoptimized + debuginfo] target(s) in 0.20s
     Running `target/debug/examples/nal-parser /Users/anatawa12/test.h264`
Input file size: 6926816
Parsing NAL took: 486.273834ms
Total NALs: 1001
$ cargo run --release --example nal-parser -- ~/test.h264
    Finished release [optimized] target(s) in 0.20s
     Running `target/release/examples/nal-parser /Users/anatawa12/test.h264`
Input file size: 6926816
Parsing NAL took: 45.588834ms
Total NALs: 1001
```